fastapi_project/requirements.txt
aerich==0.6.2
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.20.46
botocore==1.23.46
broadcaster==0.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
ciso8601==2.2.0
click==8.0.3
cPython==0.0.6
cryptography==36.0.1
databases==0.5.5
ddlparse==1.10.0
decorator==5.1.1
Deprecated==1.2.13
dictdiffer==0.9.0
Django==4.0.2
dnspython==2.2.0
docker==5.0.3
ecdsa==0.17.0
email-validator==1.1.3
et-xmlfile==1.1.0
eventlet==0.33.0
extras==1.0.0
fakeredis==1.7.1
fastapi==0.75.0
fastapi-cache2==0.1.8
fastapi-limiter==0.1.4
fastapi-login==1.7.3
fastapi-mail==1.0.6
fastapi-pagination==0.9.1
fixtures==3.0.0
frozenlist==1.3.0
future==0.18.2
google-auth==2.6.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
httpcore==0.14.7
httptools==0.3.0
httpx==0.22.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imagekitio==2.2.8
iso8601==0.1.16
itsdangerous==1.1.0
Jinja2==3.0.3
jmespath==0.10.0
limits==1.6
MarkupSafe==2.0.1
motor==2.5.1
multidict==6.0.2
numpy==1.22.1
openpyxl==3.0.9
orjson==3.6.7
orm==0.3.1
outcome==1.1.0
packaging==21.3
pandas==1.4.0
passlib==1.7.4
paywix==1.4.0
pbr==5.8.1
pendulum==2.1.2
priority==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyfcm==1.5.4
PyJWT==2.3.0
pymediainfo==5.1.0
pymongo==3.12.3
pyparsing==3.0.7
pypika-tortoise==0.1.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==5.0.2
pytz==2021.3
pytzdata==2020.1
PyVimeo==1.1.0
PyYAML==6.0
razorpay==1.2.0
redis==4.1.4
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
s3transfer==0.5.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
starlette==0.17.1
starlette-context==0.3.3
starlette-session==0.4.2
testtools==2.5.0
text-unidecode==1.3
tinydb==4.6.1
toml==0.10.2
tomlkit==0.9.2
tortoise-orm==0.18.1
trio==0.20.0
tuspy==0.2.5
typesystem==0.3.1
typing_extensions==4.1.0
urllib3==1.26.8
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.7
websocket-client==1.2.3
websockets==10.1
wrapt==1.13.3
wsproto==1.1.0
yarl==1.7.2
